existing vlan is a string, convert wanted vlan before comparing (#81)
Allow VLAN bridges to be idempotent

Managing bridges for VLAN child bridges is not currently idempotent and is being applied on each run. PR fixes and adds a test case, which fails on current main branch.
- name: add vlan bridges
  openvswitch.openvswitch.openvswitch_bridge:
    bridge: vlan5
    parent: ovs0
    vlan: 5
    state: present

Debugging the have vs want results we can see that the have vlan is a string.
root@tank:/tmp# diff debug-have-vlan17 debug-want-vlan17
1c1
< {"bridge": "vlan17", "parent": "ovs0", "vlan": "17", "fail_mode": null, "external_ids": null}
\ No newline at end of file
---
> {"bridge": "vlan17", "parent": "ovs0", "vlan": 17, "fail_mode": null, "external_ids": null, "set": null}
\ No newline at end of file

Reviewed-by: Nilashish Chakraborty <nilashishchakraborty8@gmail.com>
Reviewed-by: None <None>